Revert skinning to direct G1->24 (asset::compute_skin_matrices), add validation + non-ONNX unit tests
- Revert: main.rs + mb_probe.rs: retarget::g1_to_skin -> asset::compute_skin_matrices
  This restores the proven direct G1(34)->24-bone skinning path,
  eliminating world-space matrix interpolation shearing from retarget.rs.

- Unit tests: 3 non-ONNX tests in asset.rs prove compute_skin_matrices
  produces valid non-sheared matrices (identity, rotation, spread pose).

- Runtime validation: validate_skin_matrices() checks finite, positive det,
  translation bounds per frame.

- ONNX diagnostics: per-model timing eprintlns in MotionPipeline::new to
  identify which model hangs.

- Skeleton fixes: type mismatch in skeleton tests (usize->i32 casts),
  update test assertions for 103-bone skeleton (BONE_COUNT, finger/head offsets).

+++ src/skeleton.rs
@@ -382,8 +382,8 @@ mod tests {
 
     #[test]
     fn head_chain() {
-        assert_eq!(bone_parent(HEAD_IK), C1);
-        assert_eq!(bone_parent(HEAD_SKULL), C1);
+        assert_eq!(bone_parent(HEAD_IK), C1 as i32);
+        assert_eq!(bone_parent(HEAD_SKULL), C1 as i32);
         assert_eq!(bone_parent(HEAD_JAW), HEAD_SKULL as i32);
     }
 }
